Removed some functions from helper.js
